# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"Plan","description":"Plan Controller"},{"name":"Task","description":"Task Controller"},{"name":"User","description":"User Controller"}],"paths":{"/api/v1/plan/":{"get":{"tags":["Plan"],"summary":"Get today plan","operationId":"getDayPlanUsingGET","produces":["application/json"],"responses":{"200":{"description":"Day plan created","schema":{"$ref":"#/definitions/Plan"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"basicAuth":[]}]}},"/api/v1/plan/{date}":{"get":{"tags":["Plan"],"summary":"Get plan for given date","operationId":"getDayPlanUsingGET_2","produces":["application/json"],"parameters":[{"name":"Date","in":"path","description":"Date for plan","required":true,"type":"string","format":"date","x-example":"2020-11-19"}],"responses":{"200":{"description":"Plan created for given date","schema":{"$ref":"#/definitions/Plan"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"basicAuth":[]}]}},"/api/v1/plan/{date}/{status}":{"get":{"tags":["Plan"],"summary":"Get plan with tasks with given status for given date","operationId":"getDayPlanUsingGET_1","produces":["application/json"],"parameters":[{"name":"Date","in":"path","description":"Date for plan","required":true,"type":"string","format":"date","x-example":"2020-11-19"},{"name":"Status","in":"path","description":"Task status for plan","required":true,"type":"string","x-example":"OPEN"}],"responses":{"200":{"description":"Day plan created","schema":{"$ref":"#/definitions/Plan"}},"400":{"description":"Wrong status"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"basicAuth":[]}]}},"/api/v1/task/add":{"post":{"tags":["Task"],"summary":"Add task","operationId":"addTaskUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"taskDto","description":"Task information","required":true,"schema":{"$ref":"#/definitions/TaskDto"}}],"responses":{"200":{"description":"Task added","schema":{"$ref":"#/definitions/Task"}},"201":{"description":"Created"},"400":{"description":"Wrong task information"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"basicAuth":[]}]}},"/api/v1/task/{id}":{"get":{"tags":["Task"],"summary":"Get task information","operationId":"getTaskUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Task id","required":true,"type":"integer","format":"int64","x-example":1}],"responses":{"200":{"description":"Task found","schema":{"$ref":"#/definitions/Task"}},"400":{"description":"Wrong task id"},"401":{"description":"Unauthorized"},"403":{"description":"User have no permissions to see this task"},"404":{"description":"Task with given id not found"}},"security":[{"basicAuth":[]}]},"delete":{"tags":["Task"],"summary":"Delete task with given id","operationId":"deleteTaskUsingDELETE","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Task id","required":true,"type":"integer","format":"int64","x-example":1}],"responses":{"200":{"description":"Task deleted","schema":{"$ref":"#/definitions/Task"}},"204":{"description":"No Content"},"400":{"description":"Wrong task id"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"basicAuth":[]}]}},"/api/v1/user/":{"post":{"tags":["User"],"summary":"Creates new user","operationId":"createUserUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"appUserDto","description":"User data","required":true,"schema":{"$ref":"#/definitions/AppUserDto"}}],"responses":{"200":{"description":"User created","schema":{"$ref":"#/definitions/AppUser"}},"201":{"description":"Created"},"226":{"description":"User with login already exists"},"400":{"description":"Invalid user data"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"securityDefinitions":{"basicAuth":{"type":"basic"}},"definitions":{"AppUser":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"login":{"type":"string","example":"Kpekep"},"password":{"type":"string","example":"$2a$12$wHgf1MgL4e5J50XZD1V0G.qiwdORLVPx5kP9WMl//GZ2BUwj5DD4S"},"role":{"type":"string","example":"AUTHOR","enum":["ADMIN","AUTHOR"]}},"title":"AppUser"},"AppUserDto":{"type":"object","properties":{"login":{"type":"string","example":"Kpekep"},"password":{"type":"string","example":"123456abc"}},"title":"AppUserDto"},"Plan":{"type":"object","properties":{"empty":{"type":"boolean"},"plan":{"type":"object"}},"title":"Plan"},"Task":{"type":"object","properties":{"date":{"type":"string","format":"date","example":"2020-11-19"},"endTime":{"type":"string","format":"date","example":"00:00:00"},"id":{"type":"integer","format":"int64","example":1},"startTime":{"type":"string","format":"date","example":"00:00:00"},"taskStatus":{"type":"string","example":"OPEN","enum":["COMPLETED","OPEN"]},"title":{"type":"string","example":"Test task"}},"title":"Task"},"TaskDto":{"type":"object","properties":{"date":{"type":"string","format":"date","example":"2020-11-19"},"endTime":{"type":"string","format":"date","example":"00:00:00"},"startTime":{"type":"string","format":"date","example":"00:00:00"},"taskStatus":{"type":"string","example":"OPEN","enum":["COMPLETED","OPEN"]},"title":{"type":"string","example":"Test task"}},"title":"TaskDto"}}}